+++ src/mumble/ClientUser.cpp
@@ -316,7 +316,7 @@ bool ClientUser::isActive() {
 	if (!tLastTalkStateChange.isStarted())
 		return false;
 
-	return tLastTalkStateChange.elapsed() < Global::get().s.os.uiActiveTime * 1000000U;
+	return tLastTalkStateChange.elapsed() < static_cast< unsigned int >(Global::get().s.os.uiActiveTime) * 1000000U;
 }
 
 /* From Channel.h
